Refactor to allow different route registeries
